Introduce SLEVE in Vcode 21001 via 4 rcoefs
The choice of SLEVE is made whith usage of rcoef 3 and 4

Rename Vcode 21101 to 21001

There is no non SLEVE height coordinate. The option is rather
control by the value of the rcoef3 and 4

Absorb Vcode 21101 hyb heights SLEVE in 21001 hyb heights

Make Vcode 21001 B(nk) non flat

The fact of forcing B(nk) to 1.0 was generating spurious diag vertical
velocity at t0 in the model. This was found by running the Schaer case
with SLEVE.