The test, c_levels_all, finds an error in one level.
The current value appears to be the correct one.  I cannot see any combination
of round-off errors that can produced the expected result.  Because the expected
result comes from the model, I cannot investigate it.

I have studied this number and its source in depth.  I have checked the calculation
in float's in double's and using a spreadsheet.  I conclude that the test is in error.