Correction to c_loc for levels array not starting at (1,1,1) issue #25
If the index of the first point in the surface pressure array is not (1,1),
then the call to c_loc in BODY_F_diag_withref.hf returns the wrong address.
The 3D pressure array has the same problem if the user already allocated it
with non (1,1,1) start.
The wrong addresses are passed to the C functions which compute the diagnostic
(e.g. pressure) . The correct number of elements are computed at put in the
levels array but at the wrong place. This produced Seg faults and other nasty problems.
If the surface and level arrays starts at (1,1 there is no problem)

The actual code will still return a the levels stating at (1,1,1) even if
the pressure array start at say (-1,-1). this wil have to be correct in an other
commit.