Correct the initialization of vgrid.rec.ip1, vgrid.rec.ip2
Now only constructor_build_vert fails.  That test attempts this call:
   vgd_new(vgdid,kind=5,version=1,nk=10)
and does nothing else.  That is not a valid call, since to construct
a vgrid_5001, ptop, pref, and rcoef1 are required.  The test crashes
as it should.  Therefore, with the new interface, this is no longer
a valid test.  It has been deleted.

All 120 tests pass.  (test_vgd_free, the 121st test, fails as expected.)