solve issue #28 Vcode 2001 with old style ip1 if fst file has old style ip1
Before this change when a pressure coordinate was generated from the content of a
fst file, the ip1 encoded in the descriptor was always in new style regardless
of the fst file ip1 style. This was a problem if the ip1 in the fst file are old style.
In this case, when the user wants to compute the pressure from a ip1 list in old style
(list obtained from the fst file) vgrid could find any of these in its list.