Fine tune solution for issue #28: suppress convip error message with kind 5.
To test if ip1s are old style we were converting it to old style
with convip. However, kind 5 cannot be convert to old style and
the call to convip was giving error message poluting the user
listing. Solution is don't call convip if kind 5 since we
know this is not an old style ip1.