Add !! search with datev ip3 and etiket, call c_legacy if none of the !! fit request
The modifications were needed to make the vertical pressure profiles to work in xrec. xrec make a fstlink to the file passed in input. Therefore the search for the vertical structure has to deal we all possible mix of vcode. The addition search keys idatevm, p3 and etiket help to get the right !!. For the case of operational eta file, there are no !! in it. Normally vgrid can reconstruct the !! from P0 and PT in the file. But with mixed input (eta and hyb), if there are other !! in the file, the previous version of vgrid was looping on all !! to find the vertical structure of eta, dit not find any !! related to that and stop. Now the code calls c_legacy if the search on all !! present in the file fails.