Allow vgd_new with wild card on only one ips (ip1 or ip2) and improve option quiet
GEM model first tries to find the !! with both ip1 and ip2 in call
to vgd_new. If this fails, it tries with a wild card (-1) on ip2. This
was producing an erreur message from vgrid in GEM log since previous vgrid version
of vgrid was forcing to search of both ip1 and ip2 if any of these were not wild cards.
Now a wild card can be use on any of the ips and vgrid will not write error
message if option quiet is used.