Correction to copy in and out for non INTEL compiler in vgd_levels
For compiler not having the "is_contiguous" function there was a
problem with the copy in and out code in BODY_F_diag_withref.hf.
The statement
levels(1:ni,1:nj,1:nk) = my_levels(1:ni,1:nj,1:nk)
lead to a segmentation fault with the gnu compiler. This is why:
The range of array sfc_field will always be (1:ni,1:nj) in this function
since it is declared as a real array, even if the user calls this
function with sfc_field having a negative starting index.
The "my_levels" array will also have the same horizontal range
(1:ni,1:nj,...) since is it base on the size of sfc_field.
However, the variable levels is a pointer. Therefore its horizontal
range is known by the function and may not be (1:ni,1:nj,...).
For example it may start at a negative index. Therefore the copy in
the originale code is wrong for this case:
levels(1:ni,1:nj,1:nk) = my_levels(1:ni,1:nj,1:nk)
It must be the simpler expression
levels = my_levels
As long as "levels" and "my_levels" have the same size,
which is garantied with the checks in the code, this simple copy will
be valid for any range the user may use on pointer "levels".