Solve issue #42, ip1, A, B and C possible to get for all Vcode.
Add missing test to get all ip1, A, B and C verctors for all Vcode.
Vgrid was producing segmentation fault with vgd_get for some Vcode since
some vectors containing ip1 and A,B,C coefs informations were not
defined. The vgrid code with these missing definition is
functional since they are not used internall. But when the user
tried to access them there was a problem. This is corrected now.
Ip1, A, B and C vectors irrelevant to a given Vcode contain
zero.